refactor(python): split x402 into exact/{verify,types} + client packages
Move ExactVerifier + constants into protocols/x402/exact/verify.py and the
X402* wire TypedDicts into exact/types.py. X402Adapter (server entry) stays in
protocols/x402/__init__.py and re-exports ExactVerifier + X402_VERSION. Add
empty client/ and client/exact/ packages for the upcoming exact client. Repoint
_middleware and the x402 verifier/settle tests to the new module paths. No
behavior change.
