+++ python/src/pay_kit/protocols/x402/client/__init__.py
@@ -1,0 +1,1 @@
+"""x402 ``exact`` client: challenge parsing, payment building, auto-pay transport."""
