+++ python/src/pay_kit/protocols/x402/client/exact/__init__.py
@@ -1,0 +1,1 @@
+"""x402 ``exact`` client building blocks (payment + transport)."""
