+++ python/src/pay_kit/protocols/x402/exact/__init__.py
@@ -1,0 +1,1 @@
+"""x402 ``exact`` scheme internals: structural verifier, constants, wire types."""
